feat: localize messages interface
