+++ src/i18n/locales/en/pages/messages.ts
@@ -1,0 +1,180 @@
+export const messagesPage = {
+  messages: {
+    title: 'Messages',
+    tabs: {
+      channels: 'Channels',
+      notifications: 'Notifications',
+    },
+    channels: {
+      general: 'General',
+      osu: 'osu!',
+      taiko: 'osu!taiko',
+      fruits: 'osu!catch',
+      mania: 'osu!mania',
+    },
+    chat: {
+      placeholder: 'Type a message...',
+      send: 'Send',
+      online: 'Online',
+      offline: 'Offline',
+      typing: 'Typing...',
+      cannotSend: 'Unable to send messages',
+      you: 'You',
+    },
+    notifications: {
+      markAllRead: 'Mark all as read',
+      refresh: 'Refresh',
+      empty: 'No notifications',
+      types: {
+        friend_request: 'Friend Request',
+        team_invite: 'Team Invitation',
+        system: 'System Notification',
+      },
+    },
+    privateMessage: {
+      title: 'Private Message',
+      newMessage: 'Send private message',
+      sendTo: 'Send to',
+      close: 'Close',
+      searchPlaceholder: 'Search username...',
+      messageLabel: 'Message',
+      messagePlaceholder: 'Enter your message...',
+      send: 'Send',
+      sending: 'Sending...',
+      toasts: {
+        sent: 'Private message sent',
+        sendFailed: 'Failed to send private message',
+        searchUnsupported: 'User search is not supported yet',
+        searchFailed: 'Failed to search users',
+      },
+    },
+    errors: {
+      loadFailed: 'Failed to load messages',
+      sendFailed: 'Failed to send message',
+      connectFailed: 'Connection failed',
+    },
+    loginRequired: {
+      title: 'Sign in required',
+      description: 'Please sign in to view your messages.',
+    },
+    sidebar: {
+      title: 'Message Center',
+      tabs: {
+        channels: 'Messages',
+        notifications: 'Notifications',
+      },
+      filters: {
+        all: 'All',
+        private: 'Private',
+        team: 'Team',
+        public: 'Public',
+      },
+      actions: {
+        refreshChannels: 'Refresh channels',
+        newPrivateChat: 'New private chat',
+        refreshNotifications: 'Refresh notifications',
+        markPrivateRead: 'Mark private chats read',
+      },
+      tooltips: {
+        refreshChannels: 'Refresh channel list',
+        refreshNotifications: 'Refresh notification list',
+        markPrivateRead: 'Mark private chat messages as read',
+      },
+      states: {
+        loading: 'Loading...',
+        noChannels: 'No channels',
+        noNotifications: 'No notifications',
+      },
+      from: 'From:',
+      unknownUser: 'Unknown user',
+      markAsRead: 'Mark as read',
+      channelTypes: {
+        private: 'Private chat',
+        team: 'Team channel',
+        public: 'Public channel',
+        moderated: 'Moderated',
+      },
+      noMessage: 'No messages yet',
+      avatarAlt: 'User avatar',
+      openSidebar: 'Open channel list',
+      selectPromptTitle: 'Select a channel to start chatting',
+      selectPromptDescription: 'Choose a channel or private chat from the left to begin.',
+    },
+    toasts: {
+      loadMessagesFailed: 'Failed to load messages',
+      sendMessageFailed: 'Failed to send message',
+      noPrivateNotifications: 'No private notifications to process',
+      processResult: 'Processed: {{success}} success, {{fail}} failure',
+      refreshChannelsSuccess: 'Channel list refreshed: {{total}} channels, {{privateCount}} private',
+      refreshFailed: 'Refresh failed',
+      teamRequestMissingUser: 'Unable to fetch user information',
+      teamRequestAcceptSuccess: 'Join request accepted',
+      teamRequestRejectSuccess: 'Join request declined',
+      teamRequestActionFailed: 'Failed to {{action}} the request',
+      loadPrivateChannelFailed: 'Failed to load private channel',
+    },
+    notificationsPanel: {
+      titles: {
+        teamApplication: 'Team join request',
+        teamApplicationAccept: 'Join request accepted',
+        teamApplicationReject: 'Join request declined',
+        privateMessage: 'New private message: {{userName}}',
+        teamMessage: 'New team message: {{title}}',
+        publicMessage: 'New public channel message: {{title}}',
+        privateChannelMessage: 'New private channel message: {{title}}',
+        multiplayerMessage: 'New multiplayer message: {{title}}',
+        spectatorMessage: 'New spectator channel message: {{title}}',
+        temporaryChannelMessage: 'New temporary channel message: {{title}}',
+        groupMessage: 'New group message: {{title}}',
+        systemMessage: 'New system message: {{title}}',
+        announcementMessage: 'New announcement: {{title}}',
+      },
+      contents: {
+        teamApplication: '{{userName}} applied to join your team',
+        teamApplicationAccept: 'You have joined {{teamName}}',
+        teamApplicationReject: 'Your team application was declined',
+        privateMessage: '{{userName}}: {{message}}',
+        privateMessageTruncated: '{{userName}}: {{message}}... (more available)',
+        privateMessageWithId: 'Private message from {{userName}} (ID: {{id}})',
+        privateMessageFallback: 'Private message from {{userName}}',
+        privateMessageSourceLabel: 'From user',
+        privateMessageSourceRaw: '来自用户',
+        teamChannel: 'Team channel: {{title}}',
+        publicChannel: 'Public channel: {{title}}',
+        privateChannel: 'Private channel: {{title}}',
+        multiplayerChannel: 'Multiplayer: {{title}}',
+        spectatorChannel: 'Spectator channel: {{title}}',
+        temporaryChannel: 'Temporary channel: {{title}}',
+        groupChannel: 'Group channel: {{title}}',
+        systemChannel: 'System channel: {{title}}',
+        announcementChannel: 'Announcement channel: {{title}}',
+        genericFrom: 'From {{source}}',
+      },
+      fallbacks: {
+        teamChannel: 'Team channel',
+        teamMessage: 'Team message',
+        publicChannel: 'Public channel',
+        publicMessage: 'Public message',
+        privateChannel: 'Private channel',
+        privateMessage: 'Private message',
+        multiplayer: 'Multiplayer',
+        multiplayerMessage: 'Game message',
+        spectator: 'Spectator channel',
+        spectatorMessage: 'Spectator message',
+        temporary: 'Temporary channel',
+        temporaryMessage: 'Temporary message',
+        group: 'Group channel',
+        groupMessage: 'Group message',
+        system: 'System channel',
+        systemMessage: 'System message',
+        announcement: 'Announcement channel',
+        announcementMessage: 'Announcement message',
+        unknownSource: 'Unknown source',
+      },
+    },
+    actions: {
+      accept: 'Accept',
+      reject: 'Reject',
+    },
+  },
+} as const;
