+++ src/i18n/locales/en/pages/profile.ts
@@ -1,0 +1,217 @@
+export const profilePage = {
+  profile: {
+    errors: {
+      loadFailed: 'Unable to load profile',
+      userNotFound: 'User not found',
+      checkId: 'Please check if the user ID or username is correct',
+      tryRefresh: 'Please try refreshing the page',
+    },
+    info: {
+      title: 'Player Information',
+      globalRank: 'Global Rank',
+      countryRank: 'Country Rank',
+      underConstruction: 'Remaining data is under construction',
+    },
+    stats: {
+      rankedScore: 'Ranked Score',
+      accuracy: 'Accuracy',
+      playCount: 'Play Count',
+      totalScore: 'Total Score',
+      totalHits: 'Total Hits',
+      hitsPerPlay: 'Hits Per Play',
+      maxCombo: 'Max Combo',
+      replaysWatched: 'Replays Watched',
+      medals: 'Medals',
+      pp: 'PP',
+      playTime: 'Play Time',
+    },
+    activities: {
+      title: 'Recent Activity',
+      noActivities: 'No recent activity',
+      loadFailed: 'Failed to load user activities',
+      loadMore: 'Load More',
+      loading: 'Loading...',
+      timeAgo: {
+        justNow: 'Just now',
+        minutesAgo: '{{count}} minutes ago',
+        hoursAgo: '{{count}} hours ago',
+        daysAgo: '{{count}} days ago',
+        monthsAgo: '{{count}} months ago',
+        yearsAgo: '{{count}} years ago',
+      },
+      types: {
+        rank: {
+          prefix: 'on',
+          middle: 'achieved',
+          grade: 'rank',
+          rankPrefix: 'ranked #',
+        },
+        rankLost: {
+          prefix: 'on',
+          suffix: 'lost first place',
+        },
+        achievement: 'unlocked achievement',
+        beatmapUpload: 'uploaded beatmap',
+        beatmapUpdate: 'updated beatmap',
+        beatmapRanked: 'beatmap was ranked',
+        beatmapLoved: 'beatmap was loved',
+        beatmapRevived: 'beatmap was revived',
+        beatmapDeleted: 'beatmap was deleted',
+        usernameChange: 'changed username',
+        supportAgain: 'became a Supporter again',
+        supportFirst: 'became a Supporter for the first time',
+        supportGift: 'received Supporter as a gift',
+        activity: 'performed an activity',
+      },
+    },
+    bestScores: {
+      title: 'Best Scores',
+      noScores: 'No best scores yet',
+      loadFailed: 'Failed to load best scores',
+      loadMore: 'Load More',
+      loading: 'Loading...',
+      by: 'by',
+    },
+    friends: {
+      loading: 'Loading...',
+    },
+    userPage: {
+      title: 'Personal Introduction',
+      noContent: 'No personal introduction written yet',
+      writeButton: 'Write Personal Introduction',
+      editButton: 'Edit',
+      editTitle: 'Edit Personal Introduction',
+      placeholder: 'Write your personal introduction using BBCode...',
+      editorPlaceholder: 'Write a personal introduction for {{username}}...\n\nYou can use BBCode to format text, such as:\n[b]Bold text[/b]\n[i]Italic text[/i]\n[color=red]Colored text[/color]\n\nClick toolbar buttons or use shortcuts to quickly insert formatting.',
+      processing: 'Content is being processed...',
+      cancel: 'Cancel',
+      save: 'Save',
+      saving: 'Saving...',
+      saveSuccess: 'Personal page saved successfully!',
+      saveError: 'Save failed, please try again',
+      hasUnsavedChanges: 'You have unsaved changes',
+      confirmDiscard: 'You have unsaved changes, are you sure you want to discard editing?',
+      noEditPermission: 'You do not have permission to edit this user\'s page',
+      cannotEditPage: 'Cannot edit this page',
+      loadingEditor: 'Loading editor...',
+      editPageTitle: 'Edit Personal Page',
+      editPageSubtitle: 'Edit personal introduction for {{username}}',
+      unsavedChanges: 'Unsaved changes',
+      cancelEditTooltip: 'Cancel editing (Esc)',
+      saveShortcut: 'Ctrl+S Save',
+      supportsBBCode: 'Supports BBCode format',
+      bbcodeDescription: 'BBCode is a lightweight markup language used to format text. You can use the following tags to beautify your personal introduction:',
+      usePreview: 'Use preview function to see the final effect',
+      editCover: 'Edit cover',
+    },
+    friendActions: {
+      selfProfile: 'Own profile',
+      blocked: 'User blocked',
+      mutualFollow: 'Mutual follow - you follow each other',
+      following: 'Following - you follow this user',
+      followsYou: 'Follows you - this user follows you',
+      notFollowing: 'Not following - click to see follow options',
+      unblock: 'Unblock',
+      cancelMutualFollow: 'Cancel mutual follow',
+      unfollow: 'Unfollow',
+      block: 'Block',
+      followBack: 'Follow back (mutual follow)',
+      follow: 'Follow',
+    },
+    bbcodeEditor: {
+      placeholder: 'Enter BBCode content here...',
+      toolbar: {
+        bold: 'Bold',
+        italic: 'Italic',
+        underline: 'Underline',
+        strikethrough: 'Strikethrough',
+        color: 'Color',
+        fontSize: 'Font Size',
+        image: 'Insert Image',
+        link: 'Insert Link',
+        userProfile: 'User Profile Link',
+        email: 'Email Link',
+        youtube: 'YouTube Video',
+        audio: 'Audio',
+        imagemap: 'Image Map',
+        quote: 'Quote',
+        code: 'Code Block',
+        list: 'List',
+        box: 'Collapsible Box',
+        spoiler: 'Spoiler',
+        center: 'Center Align',
+        heading: 'Heading',
+        notice: 'Notice Box',
+        help: 'Help',
+      },
+      insertText: {
+        bold: 'Bold text',
+        italic: 'Italic text',
+        underline: 'Underlined text',
+        strikethrough: 'Strikethrough text',
+        colorText: 'Colored text',
+        text: 'text',
+        linkText: 'Link text',
+        username: 'Username',
+        emailLink: 'Email link',
+        clickToVisit: 'Click to visit website',
+        infoArea: 'This is info area',
+        quoteContent: 'Quote content',
+        codeContent: 'Code content',
+        item1: 'Item 1',
+        item2: 'Item 2',
+        title: 'Title',
+        collapsibleContent: 'Collapsible content',
+        spoilerContent: 'Spoiler content',
+        centerText: 'Centered text',
+        headingText: 'Heading text',
+        importantNotice: 'Important notice',
+      },
+      colors: {
+        text: 'color text',
+      },
+      fontSize: {
+        label: 'Font Size',
+        extraSmall: 'Extra Small',
+        small: 'Small',
+        normal: 'Normal',
+        large: 'Large',
+      },
+      help: {
+        title: 'BBCode Help',
+        button: 'Help',
+        examples: {
+          bold: 'Bold',
+          italic: 'Italic',
+          underline: 'Underline',
+          strikethrough: 'Strikethrough',
+          color: 'Color',
+          size: 'Size',
+          link: 'link',
+          text: 'text',
+          imageUrl: 'Image URL',
+          quote: 'Quote',
+          code: 'Code',
+          item1: 'Item 1',
+          item2: 'Item 2',
+        },
+      },
+      modes: {
+        edit: 'Edit',
+        preview: 'Preview',
+      },
+      preview: {
+        noContent: 'Enter content to see preview',
+        generateFailed: 'Unable to generate preview',
+        generating: 'Generating preview...',
+      },
+      validation: {
+        networkError: 'Validation failed, please check network connection',
+        validating: 'Validating',
+        syntaxCorrect: 'Syntax correct',
+        errors: ' errors',
+        syntaxErrors: 'BBCode syntax errors',
+      },
+    },
+  },
+} as const;
