+++ src/i18n/locales/en/pages/teams.ts
@@ -1,0 +1,124 @@
+export const teamsPage = {
+  teams: {
+    title: 'Team Rankings',
+    description: 'View the performance and ranking of teams',
+    tabs: {
+      teams: 'Team Rankings',
+    },
+    createTeam: 'Create Team',
+    myTeam: 'My Team',
+    joinTeam: 'Join Team',
+    viewTeam: 'View Team',
+    editTeam: 'Edit Team',
+    teamRankings: 'Team Rankings',
+    loadingTeams: 'Loading team rankings data...',
+    gameModes: {
+      osu: 'osu!',
+      taiko: 'osu!taiko',
+      fruits: 'osu!catch',
+      mania: 'osu!mania',
+    },
+    rankingTypes: {
+      performance: 'Performance',
+      score: 'Score',
+    },
+    rankingType: {
+      label: 'Ranking Type',
+    },
+    subModes: {
+      vanilla: 'Vanilla',
+      relax: 'Relax',
+      autopilot: 'Autopilot',
+    },
+    teamCard: {
+      rank: 'Rank',
+      members: 'Members',
+      captain: 'Captain',
+      averageRank: 'Average Rank',
+      totalPerformance: 'Total Performance',
+      country: 'Country',
+    },
+    pagination: {
+      previous: 'Previous',
+      next: 'Next',
+      page: 'Page {{current}} of {{total}}',
+    },
+    errors: {
+      loadFailed: 'Failed to load teams',
+      noData: 'No team data available',
+    },
+    detail: {
+      createdAt: 'Created on {{date}}',
+      members: '{{count}} members',
+      captain: 'Captain',
+      membersSectionTitle: 'Members',
+      teamMembers: 'Team Members',
+      memberCount: '({{count}} people)',
+      loading: 'Loading team details...',
+      notFound: 'Team not found',
+      notFoundDescription: 'The team information was not found',
+      backToTeams: 'Back to teams',
+      teamInfo: 'Team header information',
+      coverImage: 'Cover image',
+      teamFlag: 'Team flag',
+      teamBasicInfo: 'Team basic information',
+      teamActions: 'Team action buttons',
+      captainInfo: 'Captain information',
+      memberActions: 'Member action buttons',
+      joinTeam: 'Request to Join',
+      joining: 'Requesting...',
+      leaveTeam: 'Leave Team',
+      leaving: 'Leaving...',
+      editTeam: 'Edit Team',
+      deleteTeam: 'Delete Team',
+      kickMember: 'Kick',
+      kicking: 'Kicking...',
+      joinRequestSent: 'Join request sent, please wait for captain approval',
+      leftTeam: 'Left the team',
+      confirmLeave: 'Are you sure you want to leave this team?',
+      confirmDelete: 'Are you sure you want to delete this team? This action cannot be undone!',
+      confirmKick: 'Are you sure you want to kick {{username}}?',
+      teamDeleted: 'Team deleted',
+      memberKicked: '{{username}} has been kicked from the team',
+    },
+    create: {
+      loginRequired: 'You need to login to create or edit teams',
+      nameRequired: 'Please enter team name',
+      shortNameRequired: 'Please enter team short name',
+      assetsRequired: 'Please upload team flag and cover',
+      updateSuccess: 'Team information updated successfully',
+      createSuccess: 'Team created successfully',
+      loading: 'Loading team information...',
+      editTeam: 'Edit Team',
+      createTeam: 'Create Team',
+      editDescription: 'Modify your team information',
+      createDescription: 'Create a new team and invite friends to play together',
+      basicInfo: 'Basic Information',
+      teamName: 'Team Name',
+      teamNamePlaceholder: 'Enter team name',
+      teamShortName: 'Team Short Name',
+      teamShortNamePlaceholder: 'Enter team short name',
+      shortNameDescription: 'Short name will be displayed in rankings, recommend using 2-5 characters',
+      teamFlag: 'Team Flag',
+      selectFlag: 'Select flag',
+      flagDescription: 'Standard size: 240×120px, max 2MB, supports cropping',
+      teamCover: 'Team Cover',
+      selectCover: 'Select cover',
+      coverDescription: 'Recommended size: 1920×1280px, max 10MB, supports cropping',
+      creatingTeam: 'Creating team...',
+      memberManagement: 'Member Management',
+      leaderTransfer: 'Transfer Leadership (Optional)',
+      keepCurrentLeader: 'Keep current leader',
+      leaderTransferDescription: 'Select a new leader. If not selected, current leader will remain. After transferring leadership, you will lose management permissions.',
+      currentMembers: 'Current members ({{count}} people)',
+      currentLeader: 'Current leader',
+      willBecomeLeader: 'Will become new leader',
+      noMembersAvailable: 'No members available for selection',
+      unknown: 'Unknown',
+      cancel: 'Cancel',
+      saving: 'Saving...',
+      saveChanges: 'Save Changes',
+      required: '*',
+    },
+  },
+} as const;
