+++ src/i18n/locales/zh/pages/layout.ts
@@ -1,0 +1,5 @@
+export const layout = {
+  footer: {
+    copyright: '© 2025 g0v0.top. 为节奏游戏玩家提供最佳体验',
+  },
+} as const;
