+++ src/i18n/locales/zh/pages/messages.ts
@@ -1,0 +1,180 @@
+export const messagesPage = {
+  messages: {
+    title: '消息',
+    tabs: {
+      channels: '频道',
+      notifications: '通知',
+    },
+    channels: {
+      general: '一般',
+      osu: 'osu!',
+      taiko: 'osu!taiko',
+      fruits: 'osu!catch',
+      mania: 'osu!mania',
+    },
+    chat: {
+      placeholder: '输入消息...',
+      send: '发送',
+      online: '在线',
+      offline: '离线',
+      typing: '正在输入...',
+      cannotSend: '无法发送消息',
+      you: '你',
+    },
+    notifications: {
+      markAllRead: '全部标记为已读',
+      refresh: '刷新',
+      empty: '暂无通知',
+      types: {
+        friend_request: '好友请求',
+        team_invite: '战队邀请',
+        system: '系统通知',
+      },
+    },
+    privateMessage: {
+      title: '私信',
+      newMessage: '发送私聊',
+      sendTo: '发送给',
+      close: '关闭',
+      searchPlaceholder: '搜索用户名...',
+      messageLabel: '消息内容',
+      messagePlaceholder: '输入你的消息...',
+      send: '发送',
+      sending: '发送中...',
+      toasts: {
+        sent: '私聊已发送',
+        sendFailed: '发送私聊失败',
+        searchUnsupported: '暂不支持用户搜索功能',
+        searchFailed: '搜索用户失败',
+      },
+    },
+    errors: {
+      loadFailed: '加载消息失败',
+      sendFailed: '发送消息失败',
+      connectFailed: '连接失败',
+    },
+    loginRequired: {
+      title: '需要登录',
+      description: '请登录后查看消息',
+    },
+    sidebar: {
+      title: '消息中心',
+      tabs: {
+        channels: '消息',
+        notifications: '通知',
+      },
+      filters: {
+        all: '全部',
+        private: '私聊',
+        team: '团队',
+        public: '公共',
+      },
+      actions: {
+        refreshChannels: '刷新频道',
+        newPrivateChat: '新建私聊',
+        refreshNotifications: '刷新通知',
+        markPrivateRead: '标记私聊已读',
+      },
+      tooltips: {
+        refreshChannels: '刷新频道列表',
+        refreshNotifications: '刷新通知列表',
+        markPrivateRead: '批量标记私聊消息为已读',
+      },
+      states: {
+        loading: '加载中...',
+        noChannels: '暂无频道',
+        noNotifications: '暂无通知',
+      },
+      from: '来自:',
+      unknownUser: '未知用户',
+      markAsRead: '标记为已读',
+      channelTypes: {
+        private: '私聊',
+        team: '团队频道',
+        public: '公共频道',
+        moderated: '管制',
+      },
+      noMessage: '暂无消息',
+      avatarAlt: '用户头像',
+      openSidebar: '打开频道列表',
+      selectPromptTitle: '选择一个频道开始聊天',
+      selectPromptDescription: '从左侧选择一个频道或私聊开始对话',
+    },
+    toasts: {
+      loadMessagesFailed: '加载消息失败',
+      sendMessageFailed: '发送消息失败',
+      noPrivateNotifications: '没有需要处理的私聊通知',
+      processResult: '处理完成：成功 {{success}} 个，失败 {{fail}} 个',
+      refreshChannelsSuccess: '频道列表已刷新，共 {{total}} 个频道，其中私聊 {{privateCount}} 个',
+      refreshFailed: '刷新失败',
+      teamRequestMissingUser: '无法获取用户信息',
+      teamRequestAcceptSuccess: '已接受加入请求',
+      teamRequestRejectSuccess: '已拒绝加入请求',
+      teamRequestActionFailed: '{{action}}请求失败',
+      loadPrivateChannelFailed: '加载私聊频道失败',
+    },
+    notificationsPanel: {
+      titles: {
+        teamApplication: '团队加入申请',
+        teamApplicationAccept: '团队申请已接受',
+        teamApplicationReject: '团队申请已拒绝',
+        privateMessage: '新私聊消息：{{userName}}',
+        teamMessage: '新团队消息：{{title}}',
+        publicMessage: '新公共频道消息：{{title}}',
+        privateChannelMessage: '新私有频道消息：{{title}}',
+        multiplayerMessage: '新多人游戏消息：{{title}}',
+        spectatorMessage: '新观战频道消息：{{title}}',
+        temporaryChannelMessage: '新临时频道消息：{{title}}',
+        groupMessage: '新群组消息：{{title}}',
+        systemMessage: '新系统消息：{{title}}',
+        announcementMessage: '新公告：{{title}}',
+      },
+      contents: {
+        teamApplication: '{{userName}} 申请加入您的团队',
+        teamApplicationAccept: '您已成功加入 {{teamName}}',
+        teamApplicationReject: '您的团队申请被拒绝',
+        privateMessage: '{{userName}}：{{message}}',
+        privateMessageTruncated: '{{userName}}：{{message}}...（可能还有更多内容）',
+        privateMessageWithId: '来自 {{userName}} 的私聊消息（ID: {{id}}）',
+        privateMessageFallback: '来自 {{userName}} 的私聊消息',
+        privateMessageSourceLabel: '来自用户',
+        privateMessageSourceRaw: '来自用户',
+        teamChannel: '团队频道：{{title}}',
+        publicChannel: '公共频道：{{title}}',
+        privateChannel: '私有频道：{{title}}',
+        multiplayerChannel: '多人游戏：{{title}}',
+        spectatorChannel: '观战频道：{{title}}',
+        temporaryChannel: '临时频道：{{title}}',
+        groupChannel: '群组频道：{{title}}',
+        systemChannel: '系统频道：{{title}}',
+        announcementChannel: '公告频道：{{title}}',
+        genericFrom: '来自 {{source}}',
+      },
+      fallbacks: {
+        teamChannel: '团队频道',
+        teamMessage: '团队消息',
+        publicChannel: '公共频道',
+        publicMessage: '公共消息',
+        privateChannel: '私有频道',
+        privateMessage: '私有消息',
+        multiplayer: '多人游戏',
+        multiplayerMessage: '游戏消息',
+        spectator: '观战频道',
+        spectatorMessage: '观战消息',
+        temporary: '临时频道',
+        temporaryMessage: '临时消息',
+        group: '群组频道',
+        groupMessage: '群组消息',
+        system: '系统频道',
+        systemMessage: '系统消息',
+        announcement: '公告频道',
+        announcementMessage: '公告消息',
+        unknownSource: '未知来源',
+      },
+    },
+    actions: {
+      accept: '接受',
+      reject: '拒绝',
+    },
+  },
+} as const;
